refactor: Update role components and services
