--- packages/modules/management/src/lib/components/role/jss-forms/index.ts
@@ -1,1 +1,0 @@
-export * from './role.jss-form';
