+++ packages/modules/management/src/lib/components/role/services/index.ts
@@ -1,0 +1,1 @@
+export * from './jss-forms.service';
